dists/scripts/set_ugid_quota.sh: fixes/enhancements
Rework this script to:

1. Use cat and shell's "here document" instead of echo

2. Remove error message printing ("Can't create script file") since
   it is now printed by cat itself.

3. Use awk to prevent changing /etc/mtab if it is already fixed for
   quota on case.

4. Remove old/questionable reiserfs reference. No one knowns why it was there,
   (reiserfs quota tools? who knows).

5. Try to get device name from the file, if not found revert to supplied one.

6. Add old RedHat/chkconfig style header to init script

7. Try to use chkconfig (Red Hat / RHEL / CentOS etc.)
   and update-rc.d (Debian/Ubuntu) before reverting to the old manual
   method.

This is still a work in progress, there might be bugs.

This could fix item (3) from http://bugzilla.openvz.org/1734

+++ etc/dists/scripts/set_ugid_quota.sh
@@ -22,16 +22,22 @@ SCRIPTANAME='/etc/init.d/vzquota'
 RCDIRS="/etc/rc.d /etc"
 
 if [ -z "$MAJOR" ]; then
+	/usr/sbin/update-rc.d vzquota remove > /dev/null 2>&1
+	/sbin/chkconfig --del vzquota >/dev/null 2>&1
 	rm -f ${SCRIPTANAME} > /dev/null 2>&1
 	rm -f /etc/mtab > /dev/null 2>&1
 	ln -sf /proc/mounts /etc/mtab
 	exit 0
 fi
-echo '#!/bin/sh
+cat << EOF > ${SCRIPTANAME} || exit 1
+#!/bin/sh
+# chkconfig: 2345 10 90
+# description: prepare container to use OpenVZ 2nd-level disk quotas
+
 ### BEGIN INIT INFO
 # Provides: vzquota
-# Required-Start: $local_fs $time $syslog
-# Required-Stop: $local_fs
+# Required-Start: \$local_fs \$time \$syslog
+# Required-Stop: \$local_fs
 # Default-Start: 2 3 4 5
 # Default-Stop: 0 1 6
 # Short-Description: Start vzquota at the end of boot
@@ -39,35 +45,47 @@ echo '#!/bin/sh
 ### END INIT INFO
 
 start() {
-	[ -e "/dev/'${DEVFS}'" ] || mknod /dev/'${DEVFS}' b '$MAJOR' '$MINOR'
-	rm -f /etc/mtab >/dev/null 2>&1
-	echo "/dev/'${DEVFS}' / reiserfs rw,usrquota,grpquota 0 0" > /etc/mtab
-	mnt=`grep -v " / " /proc/mounts`
-	if [ $? = 0 ]; then
-		echo "$mnt" >> /etc/mtab
+	dev=\$(awk '(\$2 == "/") && (\$4 ~ /usrquota/) && (\$4 ~ /grpquota/) {print \$1}' /etc/mtab)
+	if test -z "\$dev"; then
+		dev="/dev/${DEVFS}"
+		rm -f /etc/mtab >/dev/null 2>&1
+		echo "/dev/${DEVFS} / ${DEVFS} rw,usrquota,grpquota 0 0" > /etc/mtab
+		grep -v " / " /proc/mounts >> /etc/mtab 2>/dev/null
+		chmod 644 /etc/mtab
 	fi
-	chmod 644 /etc/mtab
+	[ -e "\$dev" ] || mknod \$dev b $MAJOR $MINOR
 	quotaon -aug
 }
-case "$1" in
+
+case "\$1" in
   start)
 	start
 	;;
   *)
 	exit
-esac ' > ${SCRIPTANAME} || {
-	echo "Unable to create ${SCRIPTANAME}"
-	exit 1
-}
+esac
+EOF
 chmod 755 ${SCRIPTANAME}
 
-RC=
-for RC in ${RCDIRS}; do
+# Debian/Ubuntu case
+if test -x /usr/sbin/update-rc.d; then
+	/usr/sbin/update-rc.d vzquota defaults > /dev/null
+	exit
+fi
+
+# RedHat/RHEL/CentOS case
+if test -x /sbin/chkconfig; then
+	/sbin/chkconfig --add vzquota >/dev/null
+	exit
+fi
+
+# Other cases, try to put into runlevels manually
+for RC in ${RCDIRS} ""; do
 	[ -d ${RC}/rc3.d ] && break
 done
 
 if [ -z "${RC}" ]; then
-	echo "Unable to find runlevel directories"
+	echo "Unable to find init.d runlevel directories (tried $RCDIRS)"
 	exit 1
 fi
